Made explore form more modular.
