+++ client/components/Explore/ExploreForm/ExploreForm.js
@@ -1,0 +1,6 @@
+import { connect } from 'react-redux';
+import ExploreForm from './ExploreFormDumb';
+
+const mapStateToProps = ({ explore: { category } }) => ({ category });
+
+export default connect(mapStateToProps)(ExploreForm);
